Major rerfactoring and Api changes. Resolve linting hints (#14)
* improve code documentation, abstraction layers and api

* fix all mypi and linting issues

* solve all linting errors in mock nodes

* further improvements of documentation and refactoring

* Further refactoring and API improvements

* Further refactoring and API improvements

* Further refactoring and API improvements

* fix tests

* Refactoring

* Refactoring, fix tests

* Minor fixes

* Prevent deadlock when training is stopped before a valid model was created

* make _get_latest_model_files async and don't run it on threadpool

* make sure no old cookies are used

* refactoring

* simplify process executor and use async process api

* Refactor executor

* adapt backdoor controls to async api

* adapt trainer_logic to async executor api

* adapt tests to async executor api

* adapt mock_trainer_logic to new api

* adapt testing_trainer_logic to new api

* fix typing error

* minor refactoring, fix of tests

* simplify state string and do all detections

* fix test test_executor_lifecycle

* fix test test_go_to_cleanup_if_no_detections_exist

* deactivate asyncio warnings

* set request throttle to avoid Error 429

* make inference_batch_size an attribute that can be overwritten by children

* add 'model_size' to ModelInformation. Required when continuing training

Fix tests

* Improve logs. Cleanup tmp-dir after model download.

* Update data_exchanger.py

clarify docstring

* Handle file opening error

* (re)enable todo/fixme warnings in mypy

* create executable_path variable to avoid redundancy

---------

+++ learning_loop_node/trainer/trainer_logic.py
@@ -5,7 +5,7 @@ import os
 import shutil
 from abc import abstractmethod
 from datetime import datetime
-from typing import Coroutine, Dict, List, Optional
+from typing import Coroutine, List, Optional
 
 from dacite import from_dict
 
@@ -18,45 +18,47 @@ from .trainer_logic_generic import TrainerLogicGeneric
 class TrainerLogic(TrainerLogicGeneric):
 
     def __init__(self, model_format: str) -> None:
-        super().__init__(model_format)
-        self.model_format: str = model_format
-        # NOTE: String to be used in the file path for the model on the server:
-        # '/{context.organization}/projects/{context.project}/models/{model_id}/{model_format}/file'
+        """This class is the base class for all trainers that use an executor to run training processes.
+        The executor is used to run the training process in a separate process."""
 
+        super().__init__(model_format)
+        self._detection_progress: Optional[float] = None
         self._executor: Optional[Executor] = None
         self.start_training_task: Optional[Coroutine] = None
+        self.inference_batch_size = 10
+
+    # ---------------------------------------- IMPLEMENTED ABSTRACT PROPERTIES ----------------------------------------
+
+    @property
+    def detection_progress(self) -> Optional[float]:
+        return self._detection_progress
+
+    # ---------------------------------------- PROPERTIES ----------------------------------------
 
     @property
     def executor(self) -> Executor:
         assert self._executor is not None, 'executor must be set, call `run_training` first'
         return self._executor
 
-    @property
-    def hyperparameters(self) -> Optional[Dict]:
-        if self._training and self._training.data and self._training.data.hyperparameter:
-            information = {}
-            information['resolution'] = self._training.data.hyperparameter.resolution
-            information['flipRl'] = self._training.data.hyperparameter.flip_rl
-            information['flipUd'] = self._training.data.hyperparameter.flip_ud
-            return information
-        return None
+    # ---------------------------------------- IMPLEMENTED ABSTRACT MEHTODS ----------------------------------------
 
     async def _train(self) -> None:
         previous_state = TrainerState.TrainModelDownloaded
         error_key = 'run_training'
-        self._executor = Executor(self.active_training.training_folder)
-        self.active_training.training_state = TrainerState.TrainingRunning
+        self._executor = Executor(self.training.training_folder)
+        self.training.training_state = TrainerState.TrainingRunning
 
         try:
             await self._start_training()
-
             last_sync_time = datetime.now()
+
             while True:
-                if not self.executor.is_process_running():
+                await asyncio.sleep(0.1)
+                if not self.executor.is_running():
                     break
                 if (datetime.now() - last_sync_time).total_seconds() > 5:
                     last_sync_time = datetime.now()
-                    if self.get_executor_error_from_log():
+                    if self._get_executor_error_from_log():
                         break
                     self.errors.reset(error_key)
                     try:
@@ -65,65 +67,49 @@ class TrainerLogic(TrainerLogicGeneric):
                         logging.warning('CancelledError in run_training')
                         raise
                     except Exception:
-                        pass
-                else:
-                    await asyncio.sleep(0.1)
+                        logging.error('Error in sync_confusion_matrix (this error is ignored)')
 
-            error = self.get_executor_error_from_log()
-            if error:
+            if error := self._get_executor_error_from_log():
                 raise TrainingError(cause=error)
-            # TODO check if this works:
+
+            # NOTE: This is problematic, because the return code is not 0 when executor was stoppen e.g. via self.stop()
             # if self.executor.return_code != 0:
-            #     self.errors.set(error_key, f'Executor return code was {self.executor.return_code}')
-            #     raise TrainingError(cause=f'Executor return code was {self.executor.return_code}')
+            #     raise TrainingError(cause=f'Executor returned with error code: {self.executor.return_code}')
 
         except TrainingError:
-            logging.exception('Error in TrainingProcess')
-            if self.executor.is_process_running():
-                self.executor.stop()
-            self.active_training.training_state = previous_state
+            logging.exception('Exception in trainer_logic._train')
+            await self.executor.stop_and_wait()
+            self.training.training_state = previous_state
             raise
 
-    async def _start_training(self):
-        self.start_training_task = None  # NOTE: this is used i.e. by tests
-        if self.can_resume():
-            self.start_training_task = self.resume()
-        else:
-            base_model_id = self.active_training.base_model_id
-            if not is_valid_uuid4(base_model_id):  # TODO this check was done earlier!
-                assert isinstance(base_model_id, str)
-                # TODO this could be removed here and accessed via self.training.base_model_id
-                self.start_training_task = self.start_training_from_scratch(base_model_id)
-            else:
-                self.start_training_task = self.start_training()
-        await self.start_training_task
-
     async def _do_detections(self) -> None:
-        context = self.active_training.context
-        model_id = self.active_training.model_id_for_detecting
-        assert model_id, 'model_id must be set'
+        context = self.training.context
+        model_id = self.training.model_uuid_for_detecting
+        if not model_id:
+            logging.error('model_id is not set! Cannot do detections.')
+            return
         tmp_folder = f'/tmp/model_for_auto_detections_{model_id}_{self.model_format}'
 
         shutil.rmtree(tmp_folder, ignore_errors=True)
         os.makedirs(tmp_folder)
         logging.info(f'downloading detection model to {tmp_folder}')
 
-        await self.data_exchanger.download_model(tmp_folder, context, model_id, self.model_format)
+        await self.node.data_exchanger.download_model(tmp_folder, context, model_id, self.model_format)
         with open(f'{tmp_folder}/model.json', 'r') as f:
             model_information = from_dict(data_class=ModelInformation, data=json.load(f))
 
         project_folder = create_project_folder(context)
         image_folder = create_image_folder(project_folder)
-        self.data_exchanger.set_context(context)
+        self.node.data_exchanger.set_context(context)
         image_ids = []
         for state, p in zip(['inbox', 'annotate', 'review', 'complete'], [0.1, 0.2, 0.3, 0.4]):
-            self.detection_progress = p
+            self._detection_progress = p
             logging.info(f'fetching image ids of {state}')
-            new_ids = await self.data_exchanger.fetch_image_uuids(query_params=f'state={state}')
+            new_ids = await self.node.data_exchanger.fetch_image_uuids(query_params=f'state={state}')
             image_ids += new_ids
             logging.info(f'downloading {len(new_ids)} images')
-            await self.data_exchanger.download_images(new_ids, image_folder)
-        self.detection_progress = 0.42
+            await self.node.data_exchanger.download_images(new_ids, image_folder)
+        self._detection_progress = 0.42
         # await delete_corrupt_images(image_folder)
 
         images = await asyncio.get_event_loop().run_in_executor(None, images_for_ids, image_ids, image_folder)
@@ -131,19 +117,36 @@ class TrainerLogic(TrainerLogicGeneric):
             self.active_training_io.save_detections([], 0)
         num_images = len(images)
 
-        batch_size = 200
-        for idx, i in enumerate(range(0, num_images, batch_size)):
-            self.detection_progress = 0.5 + (i/num_images)*0.5
-            batch_images = images[i:i+batch_size]
+        for idx, i in enumerate(range(0, num_images, self.inference_batch_size)):
+            self._detection_progress = 0.5 + (i/num_images)*0.5
+            batch_images = images[i:i+self.inference_batch_size]
             batch_detections = await self._detect(model_information, batch_images, tmp_folder)
             self.active_training_io.save_detections(batch_detections, idx)
 
+    # ---------------------------------------- METHODS ----------------------------------------
+
+    async def _start_training(self):
+        self.start_training_task = None  # NOTE: this is used i.e. by tests
+        if self._can_resume():
+            self.start_training_task = self._resume()
+        else:
+            base_model_uuid_or_name = self.training.base_model_uuid_or_name
+            if not is_valid_uuid4(base_model_uuid_or_name):
+                self.start_training_task = self._start_training_from_scratch()
+            else:
+                self.start_training_task = self._start_training_from_base_model()
+        await self.start_training_task
+
+    # ---------------------------------------- OVERWRITTEN METHODS ----------------------------------------
+
     async def stop(self) -> None:
         """If executor is running, stop it. Else cancel training task."""
+        print('===============> stop received in trainer_logic.', flush=True)
+
         if not self.training_active:
             return
-        if self._executor and self._executor.is_process_running():
-            self.executor.stop()
+        if self._executor and self._executor.is_running():
+            await self.executor.stop_and_wait()
         elif self.training_task:
             logging.info('cancelling training task')
             if self.training_task.cancel():
@@ -152,33 +155,31 @@ class TrainerLogic(TrainerLogicGeneric):
                 except asyncio.CancelledError:
                     pass
                 logging.info('cancelled training task')
-                self.may_restart()
-
-    def get_log(self) -> str:
-        return self.executor.get_log()
+                self._may_restart()
 
     # ---------------------------------------- ABSTRACT METHODS ----------------------------------------
 
     @abstractmethod
-    async def start_training(self) -> None:
-        '''Should be used to start a training.'''
+    async def _start_training_from_base_model(self) -> None:
+        '''Should be used to start a training on executer, e.g. self.executor.start(cmd).'''
 
     @abstractmethod
-    async def start_training_from_scratch(self, base_model_id: str) -> None:
-        '''Should be used to start a training from scratch.
-        base_model_id is the id of a pretrained model provided by self.provided_pretrained_models.'''
+    async def _start_training_from_scratch(self) -> None:
+        '''Should be used to start a training from scratch on executer, e.g. self.executor.start(cmd).
+        NOTE base_model_id is now accessible via self.training.base_model_id 
+        the id of a pretrained model provided by self.provided_pretrained_models.'''
 
     @abstractmethod
-    def can_resume(self) -> bool:
+    def _can_resume(self) -> bool:
         '''Override this method to return True if the trainer can resume training.'''
 
     @abstractmethod
-    async def resume(self) -> None:
+    async def _resume(self) -> None:
         '''Is called when self.can_resume() returns True.
         One may resume the training on a previously trained model stored by self.on_model_published(basic_model).'''
 
     @abstractmethod
-    def get_executor_error_from_log(self) -> Optional[str]:
+    def _get_executor_error_from_log(self) -> Optional[str]:
         '''Should be used to provide error informations to the Learning Loop by extracting data from self.executor.get_log().'''
 
     @abstractmethod
